--- conjure-java-core-tests/src/test/resources/example-types-no-static-factory.yml
@@ -1,20 +1,0 @@
-types:
-  imports:
-    ExampleExternalReference:
-      base-type: string
-      external:
-        java: com.external.test.ExampleExternalReference
-  definitions:
-    default-package: com.palantir.product
-    objects:
-      EmptyExample:
-        fields: { }
-        docs: There are no fields in this type. A static factory method (`of`) should be generated.
-      StringExampleNoStaticFactory:
-        fields:
-          string: string
-      CovariantListExampleNoStaticFactory:
-        fields:
-          items: list<any>
-          externalItems: list<ExampleExternalReference>
-          optionalField: optional<string>
