+++ conjure-java-core/build.gradle
@@ -122,3 +122,4 @@ tasks.withType(JavaCompile) {
     // this complains about the integrationInput directory
     options.errorprone.disable 'StrictUnusedVariable'
 }
+compileJava.dependsOn compileIntegrationInputJava
