+++ conjure-java-core/src/integrationInput/java/nostaticfactory/com/palantir/product/EmptyExample.java
