+++ conjure-java-core/src/integrationInput/java/nostaticfactory/com/palantir/product/StringExampleNoStaticFactory.java
