+++ conjure-java-core/src/test/java/com/palantir/conjure/java/parameterized/FilePath.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.internal;
+package com.palantir.conjure.java.parameterized;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
