+++ settings.gradle
@@ -22,4 +22,3 @@ include 'conjure-undertow-annotations'
 include 'conjure-undertow-lib'
 include 'conjure-undertow-processor'
 include 'conjure-undertow-processor-example'
-include 'conjure-java-core-tests'
